requirements.txt
certifi==2020.4.5.1
chardet==3.0.4
cos-python-sdk-v5==1.7.9
dicttoxml==1.7.4
Django==1.11.28
django-redis==4.11.0
djangorestframework==3.11.0
idna==2.9
Pillow==7.1.1
pytz==2019.3
qcloudsms-py==0.1.4
redis==3.4.1
requests==2.23.0
six==1.14.0
urllib3==1.25.9
